Use CMake to download GoogleTest as part of build
Having GoogleTest source in project is the hardest approach to keep up
to date and some organizations may not permit this method.

This change is to use CMake to download GoogleTest as part of the
build's configuration step.

It also has the benefit of reducing the number of steps to build the unit
test as unzip is not necessary anymore.